Change price_day field type from DOUBLE to DECIMAL(13, 2) in WebTemplateHistory model for improved precision

+++ src/models/WebTemplateHistory.js
@@ -26,7 +26,7 @@ const model = (sequelize) => {
         allowNull: false,
       },
       price_day: {
-        type: DataTypes.DOUBLE,
+        type: DataTypes.DECIMAL(13, 2), // 13 digits in total, 2 after the decimal point
         allowNull: false,
       },
       subject: {
